Added transitive references between object to establish the instantiation chain. An object now knows its creator and who it has created itself.
Changed how objects are attached to CPUs when they are initially deployed. The change means that all transitive objects initially created in the context of the deployed object will be deployed to the same CPU, instead of on the vCPU. 

Added a connectToBus method to the system class, and made the necessary changes to implement the method functionality.

git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/branches/vdmj_dyn_reconfig@2276 2b2d4d09-c93c-414a-ad04-6e973f5c1cff